renamed scss imports and cleanup
