+++ src/interface/src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import 'shared/_variables.scss';
+@import 'variables';
 
 :host {
   display: flex;
