+++ src/interface/src/app/home/planning-areas/planning-areas.component.scss
@@ -1,4 +1,4 @@
-@import '../../shared/_variables.scss';
+@import 'variables';
 
 :host {
   overflow: hidden;
