+++ src/interface/src/app/home/share-plan-dialog/share-plan-dialog.component.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/mixins";
-@import "../../../styles/colors";
+@import "mixins";
+@import "colors";
 
 :host {
   display: block;
