+++ src/interface/src/app/home/welcome/welcome.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/colors";
+@import "colors";
 
 :host {
   flex: 1;
