+++ src/interface/src/app/map/region-dropdown/region-dropdown.component.scss
@@ -1,4 +1,4 @@
-@import "../../shared/_variables.scss";
+@import "variables";
 
 // Region select dropdown arrow
 @mixin arrow {
