+++ src/interface/src/app/map/stringify-map-config.pipe.ts
