+++ src/interface/src/app/plan/plan-summary/saved-scenarios/saved-scenarios.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
 @use '@angular/material' as mat;
-@import 'mixins';
+@import "mixins";
 @import 'colors';
 
 .saved-scenarios-wrapper {
