+++ src/interface/src/app/plan/plan.component.scss
@@ -1,4 +1,4 @@
-@import '../shared/_variables.scss';
+@import 'variables';
 
 .root-container {
   display: flex;
