+++ src/interface/src/app/shared/field-alert/field-alert.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/_mixins';
-@import '../../../styles/_colors';
+@import "mixins";
+@import 'colors';
 
 .field-alert-box {
   position: absolute;
