+++ src/interface/src/app/shared/form-message-box/form-message-box.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
+@import "colors";
+@import "mixins";
 
 .message-box {
   display: flex;
