+++ src/interface/src/app/shared/nav-bar/nav-bar.component.scss
@@ -1,6 +1,6 @@
-@import '../../shared/_variables.scss';
-@import "../../../styles/colors";
-@import "../../../styles/mixins";
+@import 'variables';
+@import "colors";
+@import "mixins";
 
 .nav-bar {
   display: flex;
