+++ src/interface/src/app/shared/section-loader/section-loader.component.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/mixins";
-@import "../../../styles/colors";
+@import "mixins";
+@import "colors";
 
 :host {
   display: block;
