+++ src/interface/src/app/shared/share-explore-dialog/share-explore-dialog.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/colors";
+@import "colors";
 
 :host {
   display: block;
