+++ src/interface/src/app/standalone/about/about.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
+@import "colors";
+@import "mixins";
 
 .about-container {
   gap: 30px;
