+++ src/interface/src/app/standalone/account-validation/account-validation.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
+@import "colors";
+@import "mixins";
 
 :host {
   display: flex;
