+++ src/interface/src/app/standalone/forget-password/forget-password.component.scss
@@ -1,6 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
-@import '../../../styles/forms';
+@import "colors";
+@import "mixins";
 
 :host {
   display: flex;
