+++ src/interface/src/app/standalone/horizonal-card/horizonal-card.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
+@import "colors";
+@import "mixins";
 
 .horizontal-card {
   margin-left: auto;
