+++ src/interface/src/app/standalone/info-card/info-card.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
+@import "colors";
+@import "mixins";
 
 :host {
   margin-left: auto;
@@ -18,6 +18,7 @@
   padding: 40px 40px 30px 40px;
   gap: 8px;
 }
+
 .list-icon {
   height: 20px;
   width: 20px;
