+++ src/interface/src/app/standalone/login/login.component.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/mixins';
+@import "colors";
+@import "mixins";
 
 :host {
   display: flex;
