+++ src/interface/src/styleguide/button/button.component.scss
@@ -1,5 +1,5 @@
-@import "../../styles/colors";
-@import "../../styles/mixins";
+@import "colors";
+@import "mixins";
 
 
 :host {
