+++ src/interface/src/styleguide/docs/documentation.json
@@ -133,7 +133,7 @@
             "assetsDirs": [],
             "styleUrlsData": [
                 {
-                    "data": "@import \"../../styles/colors\";\n@import \"../../styles/mixins\";\n\n\n:host {\n  box-sizing: border-box;\n  display: flex;\n  align-items: center;\n  font-size: 14px;\n  font-weight: 500;\n  line-height: 20px;\n  border-radius: 4px;\n  padding: 12px 12px;\n  cursor: pointer;\n  gap: 8px;\n  text-decoration: none;\n  height: 40px;\n  border: 1px solid transparent;\n  background-color: transparent;\n\n  &.primary-button {\n    color: $color-white;\n    background-color: $color-standard-blue;\n    border-color: $color-standard-blue;\n\n    &:active, &:hover {\n      background-color: $color-main-blue;\n      border-color: $color-main-blue;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n      border-color: $color-soft-gray;\n      background-color: $color-soft-gray;\n    }\n  }\n\n  &.ghost-button {\n    color: $color-standard-blue;\n    border-color: $color-standard-blue;\n    background-color: $color-white;\n\n\n    &:active, &:hover {\n      color: $color-white;\n      background-color: $color-main-blue;\n      border-color: $color-main-blue;\n    }\n\n\n    &:disabled {\n      color: $color-text-gray;\n      border-color: $color-soft-gray;\n      background-color: $color-soft-gray;\n    }\n  }\n\n  &.negative-button {\n    border-color: $color-error;\n    background-color: $color-error;\n    color: $color-white;\n\n    &:active, &:hover {\n      background: #B40000;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n      background-color: $color-soft-gray;\n      border-color: $color-soft-gray;\n    }\n  }\n\n  &.text-button {\n    background-color: transparent;\n    border: none;\n    color: $color-standard-blue;\n\n    &:active, &:hover {\n      background-color: $color-soft-gray;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n    }\n  }\n\n  &.positive-button {\n    border-color: $color-brand-teal;\n    background-color: $color-brand-teal;\n    color: $color-white;\n\n    &:active, &:hover {\n      background: $color-brand-green;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n      background-color: $color-soft-gray;\n      border-color: $color-soft-gray;\n    }\n  }\n\n\n  &.has-error:not(:disabled) {\n    background-color: white;\n    color: $color-error;\n    border-color: $color-error;\n  }\n\n}\n\n\n",
+                    "data": "@import \"colors\";\n@import \"mixins\";\n\n\n:host {\n  box-sizing: border-box;\n  display: flex;\n  align-items: center;\n  font-size: 14px;\n  font-weight: 500;\n  line-height: 20px;\n  border-radius: 4px;\n  padding: 12px 12px;\n  cursor: pointer;\n  gap: 8px;\n  text-decoration: none;\n  height: 40px;\n  border: 1px solid transparent;\n  background-color: transparent;\n\n  &.primary-button {\n    color: $color-white;\n    background-color: $color-standard-blue;\n    border-color: $color-standard-blue;\n\n    &:active, &:hover {\n      background-color: $color-main-blue;\n      border-color: $color-main-blue;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n      border-color: $color-soft-gray;\n      background-color: $color-soft-gray;\n    }\n  }\n\n  &.ghost-button {\n    color: $color-standard-blue;\n    border-color: $color-standard-blue;\n    background-color: $color-white;\n\n\n    &:active, &:hover {\n      color: $color-white;\n      background-color: $color-main-blue;\n      border-color: $color-main-blue;\n    }\n\n\n    &:disabled {\n      color: $color-text-gray;\n      border-color: $color-soft-gray;\n      background-color: $color-soft-gray;\n    }\n  }\n\n  &.negative-button {\n    border-color: $color-error;\n    background-color: $color-error;\n    color: $color-white;\n\n    &:active, &:hover {\n      background: #B40000;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n      background-color: $color-soft-gray;\n      border-color: $color-soft-gray;\n    }\n  }\n\n  &.text-button {\n    background-color: transparent;\n    border: none;\n    color: $color-standard-blue;\n\n    &:active, &:hover {\n      background-color: $color-soft-gray;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n    }\n  }\n\n  &.positive-button {\n    border-color: $color-brand-teal;\n    background-color: $color-brand-teal;\n    color: $color-white;\n\n    &:active, &:hover {\n      background: $color-brand-green;\n    }\n\n    &:disabled {\n      color: $color-text-gray;\n      background-color: $color-soft-gray;\n      border-color: $color-soft-gray;\n    }\n  }\n\n\n  &.has-error:not(:disabled) {\n    background-color: white;\n    color: $color-error;\n    border-color: $color-error;\n  }\n\n}\n\n\n",
                     "styleUrl": "./button.component.scss"
                 }
             ],
