+++ src/interface/src/styleguide/styles/colors.scss
@@ -1,5 +1,5 @@
 @import "../../styles/colors";
-@import "../../styles/mixins";
+@import "mixins";
 
 .color-black {
   background-color: $color-black;
