+++ src/interface/src/styleguide/styles/typography.scss
@@ -1,4 +1,4 @@
-@import "../../styles/mixins";
+@import "mixins";
 
 .regular-paragraph {
   @include regular-paragraph();
